requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.2.0
pycryptodome==3.10.1
PyJWT==1.7.1
pytz==2021.1
pyzbar==0.1.8
qrcode==6.1
requests==2.25.1
six==1.16.0
twilio==6.60.0
urllib3==1.26.6
Werkzeug==2.0.1
